apis/flask/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask == 2.2.2
psycopg2_binary == 2.9.5
python_dotenv == 0.21.0
